Enabled Android in TouchIDExample and upgraded to RN 0.49
