+++ examples/TouchIDExample/android/settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = 'TouchIDExample'
+include ':react-native-touch-id'
+project(':react-native-touch-id').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-touch-id/android')
 
 include ':app'
