+++ examples/TouchIDExample/app.json
@@ -1,0 +1,4 @@
+{
+  "name": "TouchIDExample",
+  "displayName": "TouchIDExample"
+}
