+++ examples/TouchIDExample/index.js
@@ -1,0 +1,4 @@
+import { AppRegistry } from 'react-native';
+import App from './App';
+
+AppRegistry.registerComponent('TouchIDExample', () => App);
