+++ examples/TouchIDExample/package.json
@@ -1,14 +1,13 @@
 {
-  "name": "TouchIDExample",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start"
-  },
-  "dependencies": {
-    "react": "^15.1.0",
-    "react-native": "^0.40.0",
-    "react-native-passcode-auth": "0.1.0",
-    "react-native-touch-id": "file:../.."
-  }
+	"name": "TouchIDExample",
+	"version": "0.0.1",
+	"private": true,
+	"scripts": {
+		"start": "node node_modules/react-native/local-cli/cli.js start"
+	},
+	"dependencies": {
+		"react": "16.0.0-beta.5",
+		"react-native": "0.49.5",
+	  "react-native-touch-id": "file:../.."
+	}
 }
